refactor(bundler): move shared codes into bundler-api
